requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
cohere==5.14.0
distro==1.9.0
Django==5.1.7
djangorestframework==3.15.2
fastavro==1.10.0
filelock==3.18.0
fsspec==2025.3.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
idna==3.10
jiter==0.9.0
openai==1.68.2
packaging==24.2
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.3
tokenizers==0.21.1
tqdm==4.67.1
types-requests==2.32.0.20250306
typing_extensions==4.12.2
urllib3==2.3.0
